requirements.txt
APScheduler==3.10.4
bidict==0.23.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-APScheduler==1.13.1
Flask-SocketIO==5.3.6
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-socketio==5.11.2
pytz==2024.1
simple-websocket==1.0.0
six==1.16.0
tzdata==2024.1
tzlocal==5.2
Werkzeug==3.0.3
wsproto==1.2.0
